+++ savu/test/travis/plugin_tests/reconstruction_tests/scikit_image_recon_test.py
@@ -27,12 +27,15 @@ from savu.test.travis.framework_tests.plugin_runner_test import \
 
 
 class PluginRunnerScikitImageTest(unittest.TestCase):
+    global data_file, experiment
+    data_file = '24737.nxs'
+    experiment = None
 
     def test_process(self):
-        data_file = tu.get_test_data_path('24737.nxs')
-        process_file = tu.get_test_process_path('scikit_sart_test_process.nxs')
-        run_protected_plugin_runner(tu.set_options(data_file,
-                                                   process_file=process_file))        
+        process_list = 'reconstruction/scikit_sart_test_process.nxs'
+        options = tu.initialise_options(data_file, experiment, process_list)
+        run_protected_plugin_runner(options)
+        tu.cleanup(options)
 
 if __name__ == "__main__":
     unittest.main()
