+++ savu/test/travis/process_list_tests/map3dto4d_pipeline_test.py
@@ -30,13 +30,15 @@ from savu.test.travis.framework_tests.plugin_runner_test import \
 
 
 class Map3dto4dPipelineTest(unittest.TestCase):
+    global data_file, experiment
+    data_file = '/i12_test_data/i12_test_data.nxs'
+    experiment = None
 
     def test_process_preview(self):
-        data_file = tu.get_test_data_path('/i12_test_data/i12_test_data.nxs')
-        process_file = \
-            tu.get_test_process_path('map3dto4d_pipeline_preview_test.nxs')
-        run_protected_plugin_runner(tu.set_options(data_file,
-                                                   process_file=process_file))
+        process_list = 'map3dto4d_pipeline_preview_test.nxs'
+        options = tu.initialise_options(data_file, experiment, process_list)
+        run_protected_plugin_runner(options)
+        tu.cleanup(options)               
 
 #    def test_process(self):
 #        data_file = tu.get_test_data_path('/i12_test_data/i12_test_data.nxs')
