+++ test_data/test_process_lists/multimodal/spectrum_crop_test.nxs
